requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.2
Django==4.2.7
django-filter==23.4
djangorestframework==3.14.0
drf-yasg==1.21.7
future==0.18.3
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
serializers==0.2.4
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
